Moved train options to utils
